+++ packages/tools/client-debugger/src/Logs.ts
@@ -1,0 +1,61 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+import { ConnectionState } from "@fluidframework/container-loader";
+import { IClient } from "@fluidframework/protocol-definitions";
+
+/**
+ * Base interface for data logs, associating data with a timestamp at which the data was recorded by the debugger.
+ */
+export interface LogEntry {
+    /**
+     * The time at which some data was recorded.
+     */
+    timestamp: Date;
+}
+
+/**
+ * Represents a change in some state, coupled with a timestamp.
+ *
+ * @typeParam TState - The type of state being tracked.
+ */
+export interface StateChangeLogEntry<TState> extends LogEntry {
+    /**
+     * The new state value.
+     */
+    newState: TState;
+}
+
+/**
+ * Represents a {@link @fluidframework/container-loader#ConnectionState} change.
+ */
+export interface ConnectionStateChangeLogEntry extends StateChangeLogEntry<ConnectionState> {
+    /**
+     * When associated with a new connection (i.e. state transition to
+     * {@link @fluidframework/container-loader#ConnectionState.Connected}), this will be the new client ID.
+     *
+     * Will always be undefined for disconnects.
+     */
+    clientId: string | undefined;
+}
+
+/**
+ * Represents a processed operation (op), paired with a timestamp.
+ */
+export interface AudienceChangeLogEntry extends LogEntry {
+    /**
+     * The ID of the client that was added or removed.
+     */
+    clientId: string;
+
+    /**
+     * Metadata abou the client that was added or removed.
+     */
+    client: IClient;
+
+    /**
+     * Whether the change represents a client being added to or removed from the collaborative session.
+     */
+    changeKind: "added" | "removed";
+}
